Fix AttributeNames instance for literal tuple, add test, update to izumi 0.10.10, merge envs
* Change D4SDockerModule to directly override D4SModule
