+++ d4s-test/src/main/scala/d4s/test/envs/D4SDockerModule.scala
@@ -1,20 +1,22 @@
 package d4s.test.envs
 
 import d4s.config.DynamoConfig
-import distage.TagKK
+import d4s.modules.D4SModule
+import distage.{ModuleDef, TagKK}
 import izumi.distage.config.ConfigModuleDef
 import izumi.distage.docker.examples.{DynamoDocker, DynamoDockerModule}
 import izumi.distage.model.definition.Id
 
 class D4SDockerModule[F[+_, +_]: TagKK] extends ConfigModuleDef {
-  include(new DynamoDockerModule[F[Throwable, ?]])
   include(D4SDockerModule.config)
-
-  make[DynamoConfig].from {
-    (cfg: DynamoConfig @Id("test-config"), docker: DynamoDocker.Container) =>
-      val knownAddress = docker.availablePorts.availablePorts(DynamoDocker.primaryPort).head
-      cfg.copy(endpointUrl = Some(s"http://${knownAddress.hostV4}:${knownAddress.port}"))
-  }
+  include(DynamoDockerModule[F[Throwable, ?]])
+  include(D4SModule[F] overridenBy new ModuleDef {
+    make[DynamoConfig].from {
+      (cfg: DynamoConfig @Id("test-config"), docker: DynamoDocker.Container) =>
+        val knownAddress = docker.availablePorts.availablePorts(DynamoDocker.primaryPort).head
+        cfg.copy(endpointUrl = Some(s"http://${knownAddress.hostV4}:${knownAddress.port}"))
+    }
+  })
 }
 
 object D4SDockerModule {
