+++ d4s/src/main/scala_2.12/d4s/codecs/AttributeNamesScala213.scala
@@ -1,0 +1,3 @@
+package d4s.codecs
+
+trait AttributeNamesScala213
