+++ d4s/src/main/scala_2.13/d4s/codecs/AttributeNamesScala213.scala
@@ -1,0 +1,9 @@
+package d4s.codecs
+
+trait AttributeNamesScala213 {
+
+  implicit def literalTupleAttributeNames[K <: String: ValueOf, V]: AttributeNames[(K, V)] = {
+    AttributeNames(Set[String](valueOf))
+  }
+
+}
